requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
ajax==3.0.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.2
charset-normalizer==3.1.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.1
decorators==2.0.7
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-rest-auth==3.0.0
Django==4.0.10
django-allauth==0.52.0
django-appconf==1.0.5
django-cors-headers==3.14.0
django-countries==7.5.1
django-crispy-forms==2.0
django-filter==22.1
django-node==4.0.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-react==0.11.1
django-react-tools==0.2.15
django-rest-auth==0.9.5
django-webpack==3.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==1.16.0
docker-pycreds==0.4.0
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
fabio==2022.12.1
fasteners==0.14.1
frozenlist==1.3.3
future==0.18.2
gitdb==4.0.10
GitPython==3.1.31
gpg===1.13.1-unknown
h5py==3.8.0
httplib2==0.14.0
idna==2.8
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
monotonic==1.5
multidict==6.0.4
netifaces==0.10.4
numpy==1.24.2
oauthlib==3.1.0
olefile==0.46
openai==0.27.2
openpyxl==3.1.1
optional-django==0.2.1
packaging==20.3
paramiko==2.6.0
pathtools==0.1.2
pexpect==4.6.0
phonenumbers==8.13.8
Pillow==7.0.0
protobuf==4.22.1
psutil==5.9.4
psycopg2==2.9.5
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-environ==0.4.54
python3-openid==3.2.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
react==4.3.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
reverse==0.1.0
SecretStorage==2.3.1
sentry-sdk==1.17.0
setproctitle==1.3.2
silx==1.1.2
simplejson==3.16.0
six==1.14.0
smmap==5.0.0
sqlparse==0.4.3
systemd-python==234
tqdm==4.65.0
typing-extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24
usb-creator==0.3.7
validate-email==1.3
wadllib==1.3.3
wandb==0.14.0
xkit==0.0.0
yarl==1.8.2
